refactor(config): extract server config model
