+++ crates/ecstore/Cargo.toml
@@ -43,7 +43,7 @@ rustfs-signer.workspace = true
 rustfs-storage-api.workspace = true
 rustfs-tls-runtime.workspace = true
 rustfs-checksums.workspace = true
-rustfs-config = { workspace = true, features = ["constants", "notify", "audit"] }
+rustfs-config = { workspace = true, features = ["constants", "notify", "audit", "server-config-model"] }
 rustfs-credentials = { workspace = true }
 rustfs-common.workspace = true
 rustfs-policy.workspace = true
