+++ crates/ecstore/src/config/mod.rs
@@ -24,8 +24,6 @@ pub mod storageclass;
 use crate::error::Result;
 use crate::store::ECStore;
 use com::{STORAGE_CLASS_SUB_SYS, lookup_configs, read_config_without_migrate};
-use rustfs_config::COMMENT_KEY;
-use rustfs_config::DEFAULT_DELIMITER;
 use rustfs_config::HEAL_SUB_SYS;
 use rustfs_config::audit::{
     AUDIT_AMQP_SUB_SYS, AUDIT_KAFKA_SUB_SYS, AUDIT_MQTT_SUB_SYS, AUDIT_MYSQL_SUB_SYS, AUDIT_NATS_SUB_SYS, AUDIT_POSTGRES_SUB_SYS,
@@ -36,14 +34,15 @@ use rustfs_config::notify::{
     NOTIFY_POSTGRES_SUB_SYS, NOTIFY_PULSAR_SUB_SYS, NOTIFY_REDIS_SUB_SYS, NOTIFY_WEBHOOK_SUB_SYS,
 };
 use rustfs_config::oidc::IDENTITY_OPENID_SUB_SYS;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::sync::LazyLock;
-use std::sync::{Arc, OnceLock, RwLock};
+use std::sync::{Arc, RwLock};
+
+// RUSTFS_COMPAT_TODO(CFG-004): keep old rustfs_ecstore::config model import paths while server-config model consumers migrate. Remove after all consumers import Config, KV, KVS, DEFAULT_KVS, and register_default_kvs from rustfs_config::server_config.
+pub use rustfs_config::server_config::{Config, DEFAULT_KVS, KV, KVS, register_default_kvs};
 
 pub static GLOBAL_STORAGE_CLASS: LazyLock<RwLock<storageclass::Config>> =
     LazyLock::new(|| RwLock::new(storageclass::Config::default()));
-pub static DEFAULT_KVS: LazyLock<OnceLock<HashMap<String, KVS>>> = LazyLock::new(OnceLock::new);
 pub static GLOBAL_SERVER_CONFIG: LazyLock<RwLock<Option<Config>>> = LazyLock::new(|| RwLock::new(None));
 pub static GLOBAL_CONFIG_SYS: LazyLock<ConfigSys> = LazyLock::new(ConfigSys::new);
 
@@ -98,158 +97,6 @@ pub async fn init_global_config_sys(api: Arc<ECStore>) -> Result<()> {
 
 pub async fn try_migrate_server_config(api: Arc<ECStore>) {
     com::try_migrate_server_config(api).await
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct KV {
-    pub key: String,
-    pub value: String,
-    #[serde(default, alias = "hiddenIfEmpty")]
-    pub hidden_if_empty: bool,
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct KVS(pub Vec<KV>);
-
-impl Default for KVS {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl KVS {
-    pub fn new() -> Self {
-        KVS(Vec::new())
-    }
-    pub fn get(&self, key: &str) -> String {
-        if let Some(v) = self.lookup(key) { v } else { "".to_owned() }
-    }
-    pub fn lookup(&self, key: &str) -> Option<String> {
-        for kv in self.0.iter() {
-            if kv.key.as_str() == key {
-                return Some(kv.value.clone());
-            }
-        }
-
-        None
-    }
-
-    ///Check if KVS is empty.
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    /// Returns a list of all keys for the current KVS.
-    /// If the "comment" key does not exist, it will be added.
-    pub fn keys(&self) -> Vec<String> {
-        let mut found_comment = false;
-        let mut keys: Vec<String> = self
-            .0
-            .iter()
-            .map(|kv| {
-                if kv.key == COMMENT_KEY {
-                    found_comment = true;
-                }
-                kv.key.clone()
-            })
-            .collect();
-
-        if !found_comment {
-            keys.push(COMMENT_KEY.to_owned());
-        }
-
-        keys
-    }
-
-    /// Insert or update a pair of key/values in KVS
-    pub fn insert(&mut self, key: String, value: String) {
-        for kv in self.0.iter_mut() {
-            if kv.key == key {
-                kv.value = value;
-                return;
-            }
-        }
-        self.0.push(KV {
-            key,
-            value,
-            hidden_if_empty: false,
-        });
-    }
-
-    /// Merge all entries from another KVS to the current instance
-    pub fn extend(&mut self, other: KVS) {
-        for KV { key, value, .. } in other.0.into_iter() {
-            self.insert(key, value);
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Config(pub HashMap<String, HashMap<String, KVS>>);
-
-impl Default for Config {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Config {
-    pub fn new() -> Self {
-        let mut cfg = Config(HashMap::new());
-        cfg.set_defaults();
-
-        cfg
-    }
-
-    pub fn get_value(&self, sub_sys: &str, key: &str) -> Option<KVS> {
-        if let Some(m) = self.0.get(sub_sys) {
-            m.get(key).cloned()
-        } else {
-            None
-        }
-    }
-
-    pub fn set_defaults(&mut self) {
-        if let Some(defaults) = DEFAULT_KVS.get() {
-            for (k, v) in defaults.iter() {
-                if !self.0.contains_key(k) {
-                    let mut default = HashMap::new();
-                    default.insert(DEFAULT_DELIMITER.to_owned(), v.clone());
-                    self.0.insert(k.clone(), default);
-                } else if !self.0[k].contains_key(DEFAULT_DELIMITER)
-                    && let Some(m) = self.0.get_mut(k)
-                {
-                    m.insert(DEFAULT_DELIMITER.to_owned(), v.clone());
-                }
-            }
-        }
-    }
-
-    pub fn unmarshal(data: &[u8]) -> Result<Config> {
-        let m: HashMap<String, HashMap<String, KVS>> = serde_json::from_slice(data)?;
-        let mut cfg = Config(m);
-        cfg.set_defaults();
-        Ok(cfg)
-    }
-
-    pub fn marshal(&self) -> Result<Vec<u8>> {
-        let data = serde_json::to_vec(&self.0)?;
-        Ok(data)
-    }
-
-    pub fn merge(&self) -> Config {
-        // TODO: merge default
-        self.clone()
-    }
-}
-
-pub fn register_default_kvs(kvs: HashMap<String, KVS>) {
-    let mut p = HashMap::new();
-    for (k, v) in kvs {
-        p.insert(k, v);
-    }
-
-    let _ = DEFAULT_KVS.set(p);
 }
 
 pub fn init() {
@@ -327,5 +174,24 @@ mod tests {
             .expect("heal defaults should exist");
 
         assert_eq!(heal_kvs.get(HEAL_BITROT_CYCLE), DEFAULT_HEAL_BITROT_CYCLE_SECS.to_string());
+    }
+
+    #[test]
+    fn old_config_model_path_reexports_moved_types() {
+        let mut kvs = crate::config::KVS::new();
+        kvs.insert("key".to_string(), "value".to_string());
+        let cfg = crate::config::Config(HashMap::from([(
+            "subsys".to_string(),
+            HashMap::from([(DEFAULT_DELIMITER.to_string(), kvs)]),
+        )]));
+        let moved_cfg: rustfs_config::server_config::Config = cfg;
+
+        assert_eq!(
+            moved_cfg
+                .get_value("subsys", DEFAULT_DELIMITER)
+                .expect("subsys should exist")
+                .get("key"),
+            "value"
+        );
     }
 }
